feat(map): code cleanup
