+++ src/styles/board.module.scss
