+++ src/styles/globals.scss
@@ -9,8 +9,8 @@
 }
 
 $MIN_WIDTH: 375px;
-$MAP_WIDTH: 660px;
-$MAP_HEIGHT: 600px;
+$BOARD_WIDTH: 660px;
+$BOARD_HEIGHT: 600px;
 
 // Button
 .btn-primary {
@@ -24,15 +24,15 @@ main {
   min-width: $MIN_WIDTH;
 }
 
-// Map
-.map {
+// board
+.board {
   display: block;
   position: relative;
   max-width: 660px;
   max-height: 600px;
   width: 100%;
   height: 0;
-  padding-top: min(calc($MAP_HEIGHT / $MAP_WIDTH * 100%), $MAP_HEIGHT);
+  padding-top: min(calc($BOARD_HEIGHT / $BOARD_WIDTH * 100%), $BOARD_HEIGHT);
   min-height: calc(600px / 660px * $MIN_WIDTH);
 
   .units {
