+++ test/index.html
@@ -26,13 +26,13 @@
         filter: none;
       }
 
-      .map {
+      .board {
         position: relative;
         max-width: 880px;
         height: 800px;
         margin-bottom: 50px;
       }
-      .map svg {
+      .board svg {
         width: 880px;
         height: 800px;
         position: absolute;
@@ -115,9 +115,9 @@
     </div>
 
     <div class="container">
-      <div class="map">
+      <div class="board">
         <svg>
-          <use href="../public/images/symbols.svg#map"></use>
+          <use href="../public/images/symbols.svg#board"></use>
         </svg>
       </div>
     </div>
